feat(mads): move meshmetrics generator
